+++ src/main/java/de/rwth/idsg/bikeman/ixsi/processor/PlaceAvailabilityStore.java
@@ -26,7 +26,7 @@ public class PlaceAvailabilityStore extends AbstractSubscriptionStore {
     }
 
     @Override
-    public void subscribe(String systemID, List<Long> itemIDs, long expireIntervalinMinutes) {
+    public void subscribe(String systemID, List<Long> itemIDs, Integer expireIntervalinMinutes) {
         // No op: PlaceAvailabilitySubscriptionRequestType defines no EventHorizon
     }
 }
